Break out modal views into separate files
