+++ app/javascript/components/bootcamp/modals/FinishLessonModal/FinishLessonModal.tsx
@@ -1,28 +1,15 @@
-import React, { useCallback, useMemo } from 'react'
-
+import React from 'react'
 import Modal from 'react-modal'
-import LottieAnimation from '@/components/bootcamp/common/LottieAnimation'
-import animation from '@/../animations/finish-lesson-modal-top.json'
 import { FinishLessonModalContext } from './FinishLessonModalContextWrapper'
 import { useContext } from 'react'
-import { SolveExercisePageContext } from '@/components/bootcamp/SolveExercisePage/SolveExercisePageContextWrapper'
-import useTestStore from '../../SolveExercisePage/store/testStore'
-import useTaskStore from '../../SolveExercisePage/store/taskStore/taskStore'
-// import { playAudio } from "@/utils/play-audio";
-// @ts-ignore
-// import celebrationSound from "/task-completed-sound.aac";
+import { InitialView } from '@/components/modals/welcome-modal/InitialView'
+import { CompletedBonusTasksView } from './views/CompletedBonusTasksView'
+import { CompletedExerciseView } from './views/CompletedExerciseView'
+import { CompletedLevelView } from './views/CompletedLevelView'
 
 Modal.setAppElement('body')
 export function FinishLessonModal() {
   const { isOpen } = useContext(FinishLessonModalContext)
-  // useEffect(() => {
-  //   if (!isOpen) {
-  //     return;
-  //   }
-
-  // TODO: Sort this out
-  // playAudio(celebrationSound);
-  // }, [isOpen]);
 
   return (
     // @ts-ignore
@@ -44,176 +31,9 @@ function Inner() {
       return <InitialView />
     case 'completedExercise':
       return <CompletedExerciseView />
+    case 'completedBonusTasks':
+      return <CompletedBonusTasksView />
     case 'completedLevel':
       return <CompletedLevelView />
   }
-}
-
-function InitialView() {
-  const { handleCompleteSolution, setIsOpen } = useContext(
-    FinishLessonModalContext
-  )
-
-  const { bonusTestSuiteResult } = useTestStore()
-  const { setShouldShowBonusTasks } = useTaskStore()
-
-  const hasMoreBonusTasks = useMemo(() => {
-    if (!bonusTestSuiteResult) return false
-    const uncompletedBonusTests = bonusTestSuiteResult?.tests.filter((bt) => {
-      return bt.expects.some((e) => !e.pass)
-    })
-    return uncompletedBonusTests.length > 0
-  }, [bonusTestSuiteResult])
-
-  const bonusTaskInfoText = useMemo(() => {
-    return bonusTestSuiteResult!.tests.length > 1
-      ? `There are ${
-          bonusTestSuiteResult!.tests.length
-        } bonus tasks on this exercise to complete. Do you want to go back to complete those now?`
-      : `There is a bonus task on this exercise to complete. Do you want to go back to complete it now?`
-  }, [bonusTestSuiteResult])
-
-  const handleTweakFurther = useCallback(() => {
-    setIsOpen(false)
-    if (hasMoreBonusTasks) {
-      setShouldShowBonusTasks(true)
-    }
-  }, [bonusTestSuiteResult, hasMoreBonusTasks])
-
-  return (
-    <>
-      <LottieAnimation
-        animationData={animation}
-        className="confetti"
-        style={{ height: '200px', width: '300px' }}
-      />
-      <h2 className="text-[25px] mb-12 font-semibold">Nice work!</h2>
-
-      {hasMoreBonusTasks ? (
-        <p className="text-18 leading-140 mb-8">{bonusTaskInfoText}</p>
-      ) : (
-        <p className="text-18 leading-140 mb-20">
-          You can now mark this exercise as completed and move forward, or go
-          back and carry on tweaking your code.
-        </p>
-      )}
-
-      <div className="flex items-center gap-8 self-stretch">
-        <button onClick={handleTweakFurther} className="btn-l btn-secondary">
-          {hasMoreBonusTasks ? 'Tackle bonus tasks' : 'Tweak further'}
-        </button>
-        <button
-          onClick={handleCompleteSolution}
-          className="btn-l btn-primary flex-grow"
-        >
-          Complete Exercise
-        </button>
-      </div>
-      <p className="mt-12 text-15 leading-140 text-textColor6 text-balance">
-        (If you tweak further, you can complete the exercise using the button at
-        the right when you're done)
-      </p>
-    </>
-  )
-}
-
-function CompletedExerciseView() {
-  const { nextExerciseData } = useContext(FinishLessonModalContext)
-  const { links } = useContext(SolveExercisePageContext)
-  return (
-    <div>
-      <h2 className="text-[25px] mb-12 font-semibold">Congratulations!</h2>
-
-      {nextExerciseData ? (
-        <>
-          <p className="text-18 leading-140 mb-8">
-            The next exercise is{' '}
-            <strong className="font-semibold">{nextExerciseData.title}.</strong>
-          </p>
-          <p className="text-18 leading-140 mb-20">
-            Do you want to start it now, or would you rather go back to the
-            projects list?
-          </p>
-
-          <div className="flex items-center gap-8 self-stretch">
-            <a href={links.dashboardIndex} className="btn-l btn-secondary">
-              Back to dashboard
-            </a>
-            <a
-              href={nextExerciseData.solve_url}
-              className="btn-l btn-primary flex-grow"
-            >
-              Start Next Exercise
-            </a>
-          </div>
-        </>
-      ) : (
-        <>
-          <p className="text-18 leading-140 mb-20">
-            Well done! You've finished all the exercises available to you right
-            now.
-          </p>
-
-          <div className="flex flex-col items-stretch self-stretch">
-            <a href={links.dashboardIndex} className="btn-l btn-primary">
-              Back to dashboard
-            </a>
-          </div>
-        </>
-      )}
-    </div>
-  )
-}
-
-function CompletedLevelView() {
-  const { nextLevelIdx, completedLevelIdx } = useContext(
-    FinishLessonModalContext
-  )
-  const { links } = useContext(SolveExercisePageContext)
-  return (
-    <div>
-      <h2 className="text-[25px] mb-12 font-semibold">
-        You've completed level {completedLevelIdx}!
-      </h2>
-      <p className="text-18 leading-140 mb-8">
-        <strong className="font-semibold">
-          Congratulations! That's a big achievement 🎉
-        </strong>
-      </p>
-      {nextLevelIdx ? (
-        <>
-          <p className="text-18 leading-140 mb-20">
-            You're now onto Level {nextLevelIdx} - a brand new challenge!
-            Remember to watch the teaching video in full before starting the
-            exercises.
-          </p>
-
-          <div className="flex items-center gap-8 self-stretch">
-            <a
-              href={links.bootcampLevelUrl.replace(
-                'idx',
-                nextLevelIdx.toString()
-              )}
-              className="btn-l btn-primary flex-grow"
-            >
-              Start Level {nextLevelIdx}
-            </a>
-          </div>
-        </>
-      ) : (
-        <>
-          <p className="text-18 leading-140 mb-20">
-            You've completed all the levels available to you right now. Great
-            job!
-          </p>
-
-          <div className="flex flex-col items-stretch self-stretch">
-            <a href={links.dashboardIndex} className="btn-l btn-primary">
-              Back to dashboard
-            </a>
-          </div>
-        </>
-      )}
-    </div>
-  )
 }
